+++ packages/buidler-etherscan/.env
@@ -1,5 +1,1 @@
 TS_NODE_FILES=true
-
-WALLET_PRIVATE_KEY=0xB60B1D45848F32C21EB483EE1F83EFCBE33B896E3F16044821A0DB4BF209FA8B
-
-ETHERSCAN_API_KEY=QJBGHU5IH52MIB2HNYP51ZWIBFR6MYGNY6
