+++ packages/buidler-etherscan/README.md
@@ -5,7 +5,7 @@ Buidler plugin for verifying contracts on Etherscan
 
 We recommend developing Buidler plugins using yarn. To start working on your project, just run
 
-- `yarn`
+- `npm`
 - `yarn add --peer @nomiclabs/buidler@^1.0.0-beta.2`
 
 ## Updating Buidler or other peer dependencies
