+++ packages/buidler-etherscan/package.json
@@ -2,61 +2,55 @@
   "name": "@nomiclabs/buidler-etherscan",
   "version": "0.0.1",
   "description": "Buidler TypeScript plugin for verifying contracts on etherscan",
-  "repository": "github:nomiclabs/buidler-etherscan",
+  "repository": "github:nomiclabs/buidler",
+  "homepage": "https://github.com/nomiclabs/buidler/tree/master/packages/buidler-etherscan",
   "contributors": [
     "Nomic Labs SRL",
     "Marin Petrunić <marin@nodefactory.io>"
   ],
   "license": "MIT",
-  "main": "dist/src/index.js",
-  "types": "dist/src/index.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "keywords": [
+    "ethereum",
+    "smart-contracts",
+    "buidler",
+    "buidler-plugin",
+    "etherscan"
+  ],
   "scripts": {
-    "lint:fix": "prettier --write 'src/**/*.{js,ts}' 'test/**/*.{ts}' && tslint --fix --config tslint.json --project tsconfig.json",
-    "lint": "tslint --config tslint.json --project tsconfig.json",
-    "test": "yarn run test:unit && yarn run test:integration",
-    "test:unit": "mocha --recursive \"test/unit/**/*.ts\"",
-    "test:integration": "mocha --recursive \"test/integration/**/*.ts\"",
-    "build": "tsc",
-    "buidl": "tsc",
-    "watch": "tsc -w"
+    "lint:fix": "../../node_modules/.bin/prettier --write 'src/**/*.{js,ts}' 'test/**/*.{js,ts}' && npm run lint-src -- --fix && npm run lint-tests -- --fix",
+    "lint": "npm run lint-src && npm run lint-tests",
+    "lint-tests": "../../node_modules/.bin/tslint --config tslint.json --project ./tsconfig.json",
+    "lint-src": "../../node_modules/.bin/tslint --config tslint.json --project src/tsconfig.json",
+    "test": "./scripts/run-tests.sh",
+    "build": "../../node_modules/.bin/tsc --build src",
+    "build-test": "../../node_modules/.bin/tsc --build .",
+    "clean": "rm -rf dist build-test"
   },
   "files": [
-    "dist/src/",
+    "dist/",
     "src/",
     "LICENSE",
     "README.md"
   ],
-  "devDependencies": {
-    "@types/chai": "^4.1.7",
-    "@types/download": "^6.2.4",
-    "@types/ethereumjs-abi": "^0.6.3",
-    "@types/find-up": "^2.1.1",
-    "@types/fs-extra": "^5.0.4",
-    "@types/mocha": "^5.2.5",
-    "@types/nock": "^9.3.1",
-    "@types/node": "^8.10.38",
-    "@types/request": "^2.48.1",
-    "@types/request-promise": "^4.1.42",
-    "chai": "^4.2.0",
-    "dotenv": "^6.2.0",
-    "ethers": "^4.0.27",
-    "mocha": "^5.2.0",
-    "nock": "^10.0.6",
-    "prettier": "^1.15.3",
-    "solc": "^0.5.6",
-    "source-map-support": "^0.5.9",
-    "ts-node": "^8.0.2",
-    "tslint": "^5.12.0",
-    "tslint-config-prettier": "^1.17.0",
-    "tslint-plugin-prettier": "^2.0.1",
-    "typescript": "^3.2.2"
-  },
-  "peerDependencies": {
-    "@nomiclabs/buidler": "1.0.0-beta.2"
-  },
   "dependencies": {
     "ethereumjs-abi": "^0.6.7",
     "request": "^2.88.0",
     "request-promise": "^4.2.4"
+  },
+  "devDependencies": {
+    "@nomiclabs/buidler": "1.0.0-beta.2",
+    "@types/ethereumjs-abi": "^0.6.3",
+    "@types/nock": "^9.3.1",
+    "@types/request": "^2.48.1",
+    "@types/request-promise": "^4.1.42",
+    "chai": "^4.2.0",
+    "ethers": "^4.0.27",
+    "nock": "^10.0.6",
+    "solc": "^0.5.6"
+  },
+  "peerDependencies": {
+    "@nomiclabs/buidler": "1.0.0-beta.2"
   }
 }
