+++ packages/buidler-etherscan/test/integration/PluginTests.ts
@@ -1,39 +1,27 @@
 import { TASK_FLATTEN_GET_FLATTENED_SOURCE } from "@nomiclabs/buidler/builtin-tasks/task-names";
-import { BuidlerRuntimeEnvironment } from "@nomiclabs/buidler/types";
 import { assert } from "chai";
 import { ethers } from "ethers";
 // @ts-ignore
 import * as linker from "solc/linker";
 
 import ContractCompiler from "../../src/ContractCompiler";
+import { useEnvironment } from "../helpers";
 import deployer from "../util/ContractDeployer";
-
-declare module "mocha" {
-  interface Context {
-    env: BuidlerRuntimeEnvironment;
-  }
-}
 
 describe("Plugin integration tests", function() {
   this.timeout(100000);
 
-  let env: BuidlerRuntimeEnvironment;
+  useEnvironment(__dirname + "/../buidler-project");
 
-  before("Buidler project setup", function() {
-    process.chdir(__dirname + "/../buidler-project");
-    process.env.BUIDLER_NETWORK = "develop";
-    env = require("@nomiclabs/buidler");
-  });
-
-  it("Test verifying deployed contract on etherscan", async () => {
-    let flattenedSource: string = await env.run(
+  it("Test verifying deployed contract on etherscan", async function() {
+    let flattenedSource: string = await this.env.run(
       TASK_FLATTEN_GET_FLATTENED_SOURCE
     );
 
     // need non repeating message so it makes bytecode unique
     const message = ethers.Wallet.createRandom().address;
     flattenedSource = flattenedSource.replace("placeholder", message);
-    const compilationResult = await new ContractCompiler(env.run).compile(
+    const compilationResult = await new ContractCompiler(this.env.run).compile(
       flattenedSource,
       "TestContract1"
     );
@@ -50,9 +38,10 @@ describe("Plugin integration tests", function() {
       amount
     );
     try {
-      await env.run("verify-contract", {
+      await this.env.run("verify-contract", {
         address: deployedAddress,
         contractName: "TestContract1",
+        // See: https://github.com/nomiclabs/buidler-etherscan/pull/2#discussion_r270282941
         // libraries: JSON.stringify({
         //   "SafeMath": "0x292FFB096f7221c0C879c21535058860CcA67f58"
         // }),
