+++ packages/buidler-etherscan/test/mocha.opts
@@ -1,3 +1,4 @@
 --require dotenv/config
 --require ts-node/register
 --require source-map-support/register
+--recursive test/**/*.ts
